Adding a fixer for CRASHED crawls, not incomplete.
Keep in mind that this is not the status-filling fixer.

Crashed crawls are distinguished as those that did not finish
gracefully and required an update of hits_downloaded to the actual
downloaded objects count. After the update the crawls can be found
by the usual cleaning process of the command and have agregate rows
removed.

Such crawls are a rather rare occurences - caused by deliberate killing
of the crawling process, however still a number of records did fit the
description (~2k so far).