Fixing hits_update function.
There was a major bug in the procedure when hits_consumed would be set
to 0 by subsequent runs of the function thus significantly reducing the
value calculated.

Additionally, the function will no longer update hits_consumed and
hits_posted to 0 before calculating the values. This was a major
slowdown which was only required for old data that was already
processed. Such operation should be done outside of this function.