Fixing 'correct count' by duplicate_aggregate_report.
The function will now consider hitgroups statuses related to erroneous
crawls as 'incorrect'. Additionally, it should print 'OK!' or 'ERROR!'
at the end if 'correct' count based on hitgroup statuses and hits_mv
does match.

It should be complete now, but was only tested on a subset of real data,
if 'ERROR!' is encountered, make sure that the most recent 'running'
crawl does not fall into the current interval or there is not data
processing (eg. db_arrivals) altering the data on the go.